Move allegro sync to celery task
